backend/django_app/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
Pillow>=9.0.0
python-dotenv>=1.0.0
